requirements.txt
streamlit
supabase
python-dotenv
numpy
opencv-python
pytesseract
easyocr
pillow